+++ src/Worker/RemoveContact.php
@@ -23,6 +23,7 @@ namespace Friendica\Worker;
 
 use Friendica\Core\Logger;
 use Friendica\Database\DBA;
+use Friendica\Database\DBStructure;
 use Friendica\Model\Photo;
 use Friendica\Model\Post;
 
@@ -50,7 +51,9 @@ class RemoveContact {
 			$items = Post::select(['item-id', 'post-user-id', 'uri-id', 'guid'], $condition, ['limit' => 100]);
 			while ($item = Post::fetch($items)) {
 				Logger::info('Delete removed contact item', ['id' => $item['item-id'], 'uri-id' => $item['uri-id'], 'guid' => $item['guid']]);
-				DBA::delete('item', ['id' => $item['item-id']]);
+				if (DBStructure::existsTable('item')) {
+					DBA::delete('item', ['id' => $item['item-id']]);
+				}
 				Post::delete(['uri-id' => $item['uri-id']]);
 				Post\ThreadUser::delete(['post-user-id' => $item['post-user-id']]);
 				Post\Thread::delete(['uri-id' => $item['uri-id']]);
